Moved dynamic Color etension file  in RichMessage
