+++ Sources/Utilities/UIColor+Extension.swift
@@ -112,18 +112,4 @@ public extension UIColor {
         let rgb = Int(r * 255) << 16 | Int(g * 255) << 8 | Int(b * 255) << 0
         return String(format: "%06x", rgb)
     }
-    
-    static func dynamicColor(light: UIColor, dark: UIColor) -> UIColor {
-        return UIColor { $0.userInterfaceStyle == .dark ? dark : light }
-    }
-}
-
-extension UIApplication {
-    var userInterfaceStyle: UIUserInterfaceStyle? {
-        if let windowScene = UIApplication.shared.connectedScenes
-            .first(where: { $0 is UIWindowScene }) as? UIWindowScene {
-            return windowScene.traitCollection.userInterfaceStyle
-        }
-        return nil
-    }
 }
